Refactored Context into its own file
